Services tests

See merge request Sidorenkomv/trade_back!429

+++ src/main/java/com/trade_accounting/services/impl/fias/StreetServiceImpl.java
@@ -50,7 +50,7 @@ public class StreetServiceImpl implements StreetService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Integer id) {  
         streetRepository.deleteById(id);
     }
 }
